Added an alternate way to define carriers in the faction file:   "carriers": {     "Forrestal": [       "CV-59 Forrestal",       "CV-60 Saratoga",       "CV-61 Ranger",       "CV-62 Independence"     ],     "Stennis": [       "CVN-71 Theodore Roosevelt",       "CVN-72 Abraham Lincoln",       "CVN-73 George Washington",       "CVN-74 John C. Stennis",       "CVN-75 Harry S. Truman"     ]   },
The carrier names are picked based on the carrier type. Several carriers can be included in the same campaign with names coming from different pools. Also prevented duplicate names in campaigns with more than one carrier.

If the carriers are not found in the faction file with this new format, Liberation falls back to the previous method. This retains compatibility with existing faction files.

Partially resolves #1699