Implemented the suggested change by DanAlbert:
This could be nowhere remotely near the BAI target area. I don't think it's worth fighting all the edge cases in this when we could instead just spawn at the origin point. The convoys just barely leave their origin in the span of a mission anyway, so it's a more accurate representation to have them there than 1/3 of the way between bases anyway.

I don't think the code here needs to be any more complicated than an if around the existing add_waypoint.