Several campaign designers have expressed interest in having the autoplanner also plan e.g. helicopters for CAS missions or otherwise have more variety in the airframes which the autoplanner (or the opposing force) will purchase and plan missions for. This is an attempt to provide that power. By specifying:
airbases:
  # Airbase number 2
  2:
    increased_aircraft_autoplanner_variety: true

in the campaign yaml, the autoplanner is told to ignore the priority lists for missions planned for aircraft taking off from this airbase. This will result in the autoplanner planning missions for each airframe at this base evenly for each task type. See #41 for discussion on the autoplanner behavior.