Add Option to define Presets from the Campaign.miz
This will allow campaign designers to use trigger zones to specify precisely which template or UnitGroup should be placed at this position. Can be used if randomization is not wanted for the specific place. Example you want to position a RL SA-2 Site exactly where it stands IRL with the correct heading and template.

- added PresetLocation class which will allow to parse information from campaign.miz to liberation.
- Removed the category from buildings and added a tasking for each category instead.
- Allow Task, Template, UnitGroup and Heading as properties