Split Flight into scheduled and unscheduled states.
To allow flight plans to change after the turn has started playing we
need to replace the timedelta based TOTs with datetime. For that we need
to make sure all the times are precomputed (or at least relative to an
accessible precomputed value) because it would not be practical or
performant to update every timedelta in the game each tick.

We can't practically replace the current timedeltas with datetimes
because we lack a useful zero value like we have for timedelta (custom
flight plans that aren't able to reason about their flight plan always
start immediately, with a start time of `timedelta()`). We could use
`None`, but that would require anything that needs to _resolve_ that
time to have the mission time available.

Instead, we'll just remove all the timing from the stages of the game
that don't need that information. The main problem with pre-computing
times is doing so for flights and packages that have not been fully
planned. By separating the draft state, we can use that for the corners
that don't need timing information and can compute timing information
once the flight is sufficiently planned.

TODO: Actually remove ScheduledFlight from the early phases.

https://github.com/dcs-liberation/dcs_liberation/issues/1680