Implemented GroundForcePainter for applying custom liveries to ground units. The liveries can be applied to front line objects (including infantry), convoys and TGOs.
Also implemented AircraftPainterJtac for applying a livery from the faction to generated JTAC air units.

This is only a draft for now, since it's getting the ground force liveries from the faction file. This has already been discussed on #dev and will not be the final way to find out the liveries.