Updated US Aim-9 Weapons Info
Add Aim-9B's as fallbacks for F-4s and A-4s. Rework path by capability and not year (i.e. Aim-9X>M>L>P5>P>B), leave note in P that the year can be pushed back to represent earlier models the P is based on that are missing from game.