Update f4.py
There were a number of weapons changes for the F-4. This is a direct paste from planes.py with the new weapons injected above it, I did not bother reformatting each line like I did last time, this seems much easier to read. If that is incorrect and I need to reformat it let me know and I will do so.