Updated the F4.py pydcs extension to match the updated F-4B/C mod and rework the standard payloads to add <clean> to new pylons 11 and 12.
This includes a workaround to allow Liberation to use the new VSN weapons with combined 2x Aim-9js on pylons 3 and 9 underslung with bombs on ters on new pylons 11 and 12. In mission editor the combined weapons are selected in pylons 3 and 9 and their under-slung counterparts are forced onto 11 and 12 using "required" arguments in the mod's lua. All other pylon 3 and 9 weapons use "required <clean>" arguments. Liberation doesn't have a way to force these linkages onto pylons 11 and 12 and without them, even without <clean>, no weapons will load on 3 and 9 or 11 and 12.

The workaround for normal weapons is to set the standard load-outs to <clean> on pylons 11 and 12. This allows all normal weapons to work on pylons 3 and 9 so long as pylons 11 and 12 are left as Clean. It also allows Clean into the Liberation dropdown so it can be selected. The workaround for the 4 new weapons that combines pylons 3 with 11 and 9 with 12 is that the user has to use the matching pair on each set of pylons. For example - if F4B_LAU105_AIM9J_2_BRU42A_MK82_3 is selected for pylon 3, BRU 42A MK82*3 LAU105 AIM9J*2 must be selected for pylon 11. Failure to do this correctly doesn't crash liberation or DCS, the result will just be either no weapons at all on either pylon or the underslung weapons on 11 and 12 floating without a pylon attaching it to the plane.

When updating f4.py in the future, note that running the pydcs database export doesn't pull any data for Pylons 11 and 12. Those matching weapons / classes have to be manually defined in those pylons. This is noted in f4.py.