WIP: Turns mods into mods.
DCS mods currently need explicit support in Liberation because pydcs
needs the Python classes to work with. This proof of concept works
around that by loading python code directly from a mod directory.

I'm not sure if this is a good idea. Most players probably don't
understand the security risk of allowing any mod someone posts to
discord run arbitrary code on their machine.

It also requires that we never again serialize a DCS object of a type
defined by a mod class (such as an aircraft), because those can't be
pickled. I've worked around that here because the DCS type doesn't
actually need to be serialized to UnitType, so anything that works with
UnitType (which should be everything) won't have that problem, but it's
a potential risk.